requirements.txt
numpy
pandas
scikit-learn
matplotlib
torch>=2.0.0; platform_system != "Darwin" or platform_machine != "arm64"
torch==2.3.0; platform_system == "Darwin" and platform_machine == "arm64"
joblib
typer>=0.12.0
pyyaml